+++ services/trade_producer/src/kraken_api.py
@@ -6,10 +6,12 @@ from websocket import create_connection
 
 
 class KrakenWebsocketTradeAPI:
-    def __init__(self,
-                 #product_id: str,
-                 product_id: list,
-                 URL: str = 'wss://ws.kraken.com') -> None:
+    def __init__(
+        self,
+        # product_id: str,
+        product_id: list,
+        URL: str = 'wss://ws.kraken.com',
+    ) -> None:
         self.product_id = product_id
         self.URL = URL
 
@@ -20,17 +22,33 @@ class KrakenWebsocketTradeAPI:
         # Subscribing to the trades for the given product
         self._subscribe(self.product_id)
 
-    def _subscribe(self,
-                   #product_id: str
-                   product_id: list) -> None:
+    def _subscribe(
+        self,
+        # product_id: str
+        product_id: list,
+    ) -> None:
         logger.info(f'Subscribing to trades for {product_id}')
 
         # Let's subscribe to the trades
         msg = {
             'method': 'subscribe',
             #'params': {'channel': 'trade', 'symbol': [product_id], 'snapshot': False},
-            'params': {'channel': 'trade', 'symbol': ['ETC/USD', 'ETH/USD', 'LTC/USD', 'MLN/USD', 'REP/USD', \
-                        'BTC/USD', 'XLM/USD', 'XMR/USD', 'XRP/USD', 'ZEC/USD'], 'snapshot': False},
+            'params': {
+                'channel': 'trade',
+                'symbol': [
+                    'ETC/USD',
+                    'ETH/USD',
+                    'LTC/USD',
+                    'MLN/USD',
+                    'REP/USD',
+                    'BTC/USD',
+                    'XLM/USD',
+                    'XMR/USD',
+                    'XRP/USD',
+                    'ZEC/USD',
+                ],
+                'snapshot': False,
+            },
         }
 
         logger.info(f'Sending message: {json.dumps(msg)}')
@@ -66,10 +84,10 @@ class KrakenWebsocketTradeAPI:
         if 'data' not in message:
             return []
 
-        #logger.info("Message received: ", message)
+        # logger.info("Message received: ", message)
 
         message = json.loads(message)
-        #logger.info(message)
+        # logger.info(message)
 
         # Extract trades from the message['data']
         trades = []
