+++ services/trade_producer/src/main.py
@@ -1,8 +1,36 @@
+import requests
 from loguru import logger
 from quixstreams import Application
 
 from src import config
 from src.kraken_api import KrakenWebsocketTradeAPI
+
+
+# Creates a Kraken product_id from the first_part and second_part of the assets pairs in case of Bitcoin, since
+# in Kraken the first_part is always 'XBT' for Bitcoin. Otherwise, it takes the wsname from the asset_pairs_url
+def create_kraken_pair(first_part, second_part, wsname):
+    if first_part == 'XBT':
+        base = 'BTC'
+        return f'{base}/{second_part}'
+    else:
+        return f'{ (wsname) }'
+
+
+# Retrieves a list of product_ids (assets pairs) from the specified asset_pairs_url.
+def get_crypto_pairs(asset_pairs_kraken_url: str):
+    pairs = requests.get(asset_pairs_kraken_url).json()['result']
+    crypto_pairs = []
+    parts_length = 4
+
+    for pair, data in pairs.items():
+        if pair[0] == 'X':
+            first_part, second_part = pair[:parts_length], pair[parts_length:]
+            if second_part == 'ZUSD':
+                crypto_pairs.append(
+                    create_kraken_pair(first_part[1:], second_part[1:], data['wsname'])
+                )
+
+    return crypto_pairs
 
 
 def produce_trades(
@@ -54,6 +82,6 @@ if __name__ == '__main__':
     produce_trades(
         kafka_broker_address=config.kafka_broker_address,
         kafka_topic_name=config.kafka_topic_name,
-        product_id=config.product_id,
+        product_id=get_crypto_pairs(config.asset_pairs_kraken_url),
         URL=config.URL_KRAKEN,
     )
